+++ Bolt-Core/src/Graphics/Buffers/VertexBuffer.cpp
@@ -5,12 +5,12 @@
 namespace Bolt
 {
 
-	VertexBuffer::VertexBuffer(size_t size, const BufferLayout& layout, BufferUsage usage) : VertexBuffer(nullptr, size, layout, usage)
+	VertexBuffer::VertexBuffer(uint32_t size, const BufferLayout& layout, BufferUsage usage) : VertexBuffer(nullptr, size, layout, usage)
 	{
 	
 	}
 
-	VertexBuffer::VertexBuffer(const void* data, size_t size, const BufferLayout& layout, BufferUsage usage) : Buffer(data, size, BufferTarget::ArrayBuffer, usage),
+	VertexBuffer::VertexBuffer(const void* data, uint32_t size, const BufferLayout& layout, BufferUsage usage) : Buffer(data, size, BufferTarget::ArrayBuffer, usage),
 		m_Layout(layout)
 	{
 		BLT_ASSERT(Size() % layout.Stride() == 0, "Buffer could not fit integer number of vertices");
